refactor: consolidate DrawerSwitch and FilterSwitch into AppSwitch
Replace two near-identical switch components with a single AppSwitch
that uses AppSwitchStyle enum (prominent/subtle) for visual variants.
